+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+- [#15](https://github.com/nubank/vessel/pull/15): added the user option
+  to set the default user image
 
 ## [0.2.128] - 2020-05-28